+++ .umirc.ts
@@ -1,4 +1,11 @@
 export default {
   mfsu: false,
-  plugins: ['@umijs/plugin-docs'],
+  analytics: {
+    ga: 'G-ET6VJN7XQ8',
+    baidu: '5a66c03cb0ae986f876184554f2b9e13',
+  },
+  plugins: [
+    '@umijs/plugin-docs',
+    require.resolve('./packages/plugins/src/analytics'),
+  ],
 };