+++ examples/tabs-routes/.umirc.ts
@@ -1,0 +1,15 @@
+import routes from './src/routes';
+
+export default {
+  appType: 'pc',
+  keepalive: [],
+  aconsole: {
+    console: {},
+    inspx: {},
+  },
+  mfsu: {},
+  antd: {},
+  hash: false,
+  tabsLayout: {},
+  routes,
+};