+++ packages/plugins/src/analytics.ts
@@ -1,0 +1,3 @@
+import analytics from '@umijs/plugins/dist/analytics';
+
+export default analytics;